introduce line cache for the vertex
